Time: 1693 ms (5.18%), Space: 157.8 MB (5.45%) - LeetHub

+++ 1220-count-vowels-permutation/1220-count-vowels-permutation.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
 #define MOD 1000000007
 unordered_map<string , int>m;
-long help(int n , char i )
+long help(long n , char i )
 {
     if(n==0)
     {
